TASK: Restructure Classes
