+++ Classes/Domain/TemplateFactory/StopBuildingTemplatePartException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Flowpack\NodeTemplates\Domain;
+namespace Flowpack\NodeTemplates\Domain\TemplateFactory;
 
 /** @internal implementation detail of {@see TemplateFactory} */
 class StopBuildingTemplatePartException extends \RuntimeException
