+++ Classes/Infrastructure/NodeTemplateDumper/Comment.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Flowpack\NodeTemplates\NodeTemplateDumper;
+namespace Flowpack\NodeTemplates\Infrastructure\NodeTemplateDumper;
 
 use Neos\Flow\Annotations as Flow;
